Linearity of regressed freq against mirror freq
This commit occurs after discussion of the project with Christian
regarding the direction of the project, where certain discussions such
as identiftying the source of noise should be preferred. This commit
serves to collect data for the presentation, but also shows 2pi jumps
that occur unexpectedly, and also in the abscence of the quantum
computer.