Update kwargs for updated Matplotlib requirements
ScriptLibrary was runned on Py 3.10 to have recognised broken calls
to LaTeX. Fix was recognised to have been due to use of 'useTex' for
old versions of MPL, whereas for 'usetex' on newer versions. Other
required fix includes installation of divpng package on LaTeX,
ghostscript and correct path environment in Windows path variable.