Update convolution algorithm and dependencies
The use of a for loop within a iterator for numpy foriter
proved to be rebuilding the wheel, when np.forconvolve was
identified to have done the same for much faster. Consequently,
it was discovered that the older for loop was missing out the
final data point that was supposed to be generated, and
therefore, all relevant code has been refactored.

Refactor IQ_demod functions

Older unused code from the era of 2pi corrections has since been
shifted into the archive for better keeping.